# [CPU] SHM based allreduce improvement for small message size