--- a/csrc/cpu/comm/shm.cpp
+++ b/csrc/cpu/comm/shm.cpp
@@ -520,14 +520,39 @@ static void parallel_memcpy(void* to, void* from, size_t n_bytes)
 
 size_t slice_size(size_t chunk_el, int slice_idx);
 char* slice_data(char* data_ptr, size_t chunk_el, int el_size, int slice_idx);
-static void parallel_multi_memcpy(int world_size, char *data_ptr, size_t chunk_el, size_t chunk_size, int data_size)
+static void parallel_multi_memcpy(int world_size,
+                                  char* data_ptr,
+                                  size_t chunk_el,
+                                  size_t chunk_size,
+                                  int data_size) __attribute__((target("avx512bw")));
+static void parallel_multi_memcpy(int world_size,
+                                  char* data_ptr,
+                                  size_t chunk_el,
+                                  size_t chunk_size,
+                                  int data_size)
 {
-    for (int i = 0; i < world_size; i++) {
-        int rank = (i + world_rank) % world_size;
+#if 1
+    for (int rank = 0; rank < world_size; rank++) {
         parallel_memcpy(slice_data(data_ptr, chunk_el, data_size, rank),
-                        slice_data(workspace[rank]->buffer, chunk_el, chunk_size / chunk_el, rank),
+                        slice_data(workspace[rank]->buffer, chunk_el, data_size, rank),
                         slice_size(chunk_el, rank) * data_size);
     }
+#else
+    // assuming the first slice has smaller size than any other slice
+    auto n_bytes = slice_size(chunk_el, 0) * data_size;
+    auto aligned_bytes = n_bytes - (n_bytes % VECTOR_LENGTH_IN_BYTES);
+    int step = chunk_el / world_size * data_size;
+#pragma omp parallel for
+    for (int i = 0; i < aligned_bytes; i += VECTOR_LENGTH_IN_BYTES) {
+        auto to = data_ptr;
+        for (int rank = 0; rank < world_size; rank++) {
+            auto from = workspace[rank]->buffer + step * rank;
+            auto val = _mm256_loadu_si256((__m256i*)((char*)from + i));
+            _mm256_storeu_si256((__m256i*)((char*)to + i), val);
+            to += step;
+        }
+    }
+#endif
 }
 
 #define positive_mod(num, mod) ((((num) % (mod)) + (mod)) % (mod))
@@ -692,11 +717,11 @@ void distributed_naive_reduce(char* data_ptr,
         wait_buffer_state_until_range(rank, coll_allreduce_naive__reduce_done, 2);
     }
     parallel_multi_memcpy(world_size, data_ptr, chunk_el, chunk_size, data_size);
-    //for (int i = 0; i < world_size; i++) {
-        //int rank = (i + world_rank) % world_size;
-        //parallel_memcpy(slice_data(data_ptr, chunk_el, data_size, rank),
-                        //slice_data(workspace[rank]->buffer, chunk_el, chunk_size / chunk_el, rank),
-                        //slice_size(chunk_el, rank) * data_size);
+    // for (int i = 0; i < world_size; i++) {
+    // int rank = (i + world_rank) % world_size;
+    // parallel_memcpy(slice_data(data_ptr, chunk_el, data_size, rank),
+    // slice_data(workspace[rank]->buffer, chunk_el, chunk_size / chunk_el, rank),
+    // slice_size(chunk_el, rank) * data_size);
     //}
     std::atomic_thread_fence(std::memory_order_release);
     workspace[world_rank]->state = coll_allreduce_naive__copy_out_done;