--- a/csrc/cpu/comm/shm.cpp
+++ b/csrc/cpu/comm/shm.cpp
@@ -12,7 +12,7 @@
 #include <sys/mman.h>
 #include "shm.h"
 
-#define DO_PROFILE
+//#define DO_PROFILE
 #ifdef DO_PROFILE
 #include <cfloat>
 #include <chrono>