@delock 2024-06-12
Hi @adk9 , code updated according to comments, thanks!