@adk9 2024-06-12
> Hi @adk9 , code updated according to comments, thanks!

Hi @delock, thanks for your changes! The formatting check seems to fail for this PR. Could you run `pre-commit` or `clang-format` on your changes?