@delock 2024-06-12
> > Hi @adk9 , code updated according to comments, thanks!
> 
> Hi @delock, thanks for your changes! The formatting check seems to fail for this PR. Could you run `pre-commit` or `clang-format` on your changes?

Hi @adk9 , formatting had been fixed in latest CI.  Thanks!